Adding in a refactor to make it so the fucking search filter is independent of what uses it.
