+++ src/netrunner/search.ng.html
@@ -1,0 +1,132 @@
+<div class="search-cmp container-fluid">
+  <h3> Don't Cheat on Dat </h3>
+
+  <form class="netrunner-search" #cardSearch="ngForm">
+    <span class="input-group">
+      <input style="width: 80%;" class="form-control" type="text" name="searchText" 
+       title="Filter out netrunner cards, figure out what the fuck is legal"
+       placeholder="Search for a netrunner Cards"
+       [ngModel]="searchText"
+       (ngModelChange)="searchCards($event)"
+      >
+      <span class="error" *ngIf="errMsg">{{errMsg}}</span>
+    </span>
+    <label>
+      <input name="searchFullText" type="radio" [value]="false"  [(ngModel)]="searchFullText">
+      Title Search
+    </label>
+    <label>
+      <input name="searchFullText" type="radio" [value]="true"  [(ngModel)]="searchFullText">
+      Full Text
+    </label>
+  </form>
+
+
+  <mat-expansion-panel>
+    <mat-expansion-panel-header>
+      <mat-panel-title> <h3>Factions & Types</h3> </mat-panel-title>
+      <mat-panel-description>
+        <button mat-raised-button type="button" color="primary"
+          (click)="factionToggle($event, 'Corp')"
+        >Corp</button>
+        <button mat-raised-button type="button" color="warn"
+          (click)="factionToggle($event, 'Runner')"
+        >Runner</button>
+      </mat-panel-description>
+    </mat-expansion-panel-header>
+    <button mat-raised-button type="button" accent="primary"
+      (click)="toggleSettings([factionSelection, typeSelection], true)"
+    >All</button>
+    <button mat-raised-button type="button" 
+      (click)="toggleSettings([factionSelection, typeSelection], false)"
+    >None</button>
+
+    <div>
+      CORP | Factions | Types
+      <div class="corp-factions">
+        <span *ngFor="let faction of corpFactions">
+          <mat-checkbox class="corp-faction-checkbox" 
+            [(ngModel)]="factionSelection[faction.code]"
+            (click)="toggleFaction(faction)"
+          >{{ faction.name }}</mat-checkbox> 
+        </span>
+      </div>
+      <div class="corp-types">
+        <span *ngFor="let type of corpTypes">
+          <mat-checkbox class="corp-faction-checkbox" 
+            [(ngModel)]="typeSelection[type.code]"
+            (click)="toggleType(type)"
+          >{{ type.name }}</mat-checkbox>
+        </span>
+      </div>
+    </div>
+
+    <div>
+      Runner 
+      <div class="runner-factions"> 
+        <span *ngFor="let faction of runnerFactions">
+           <mat-checkbox class="runner-faction-checkbox"
+             [(ngModel)]="factionSelection[faction.code]"
+             (click)="toggleFaction(faction)"
+          >{{ faction.name }}</mat-checkbox>
+        </span>
+      </div>
+      <div class="runner-types">
+        <span *ngFor="let type of runnerTypes">
+           <mat-checkbox class="runner-faction-checkbox"
+            [(ngModel)]="typeSelection[type.code]"
+            (click)="toggleType(type)"
+          >{{ type.name }}</mat-checkbox>
+        </span>
+      </div>
+    </div>
+  </mat-expansion-panel>
+
+  <mat-expansion-panel>
+    <mat-expansion-panel-header>
+      <mat-panel-title> <h3>Cycle Selection</h3> </mat-panel-title>
+      <mat-panel-description>
+        <button mat-raised-button type="button" 
+          color="accent"
+          (click)="rotationStateChange(rotationState, $event)">
+          {{rotationStates[rotationState]}}
+        </button>
+      </mat-panel-description>
+    </mat-expansion-panel-header>
+
+    <div class="nr-search-filter">
+      <button mat-raised-button type="button" 
+        color="primary"
+        (click)="toggleSettings([cycleSelection, packSelection], true, $event)"
+      >All</button>
+      <button mat-raised-button type="button" 
+        (click)="toggleSettings([cycleSelection, packSelection], false, $event)"
+      >None</button>
+      <button mat-raised-button type="button" 
+        color="accent"
+        (click)="rotationStateChange(rotationState, $event)"
+      >{{rotationStates[rotationState]}}</button>
+      <div class="cycle-selection">
+        <div *ngIf="cycles">
+          <ul *ngFor="let cycle of cycles">
+            <mat-checkbox class="cycle-selection"
+              [(ngModel)]="cycleSelection[cycle.code]" 
+              (change)="toggleCycle(cycle)"
+            >{{cycle.name}} : ({{cycle.size}}) Rotated: {{cycle.rotated}}</mat-checkbox>
+              <li *ngFor="let pack of (packListing[cycle.code] || [])">
+                <mat-checkbox class="pack-selection"
+                  [(ngModel)]="packSelection[pack.code]"
+                  (change)="togglePack(pack)"
+                >{{pack.name}} - {{pack.date_release}} : ({{pack.size}})</mat-checkbox>
+              </li>
+          </ul>
+        </div>
+      </div>
+    </div>
+  </mat-expansion-panel>
+
+
+  <div>
+    Total Cards: {{allCards?.length}}
+  </div>
+</div>
